+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 2.0.0
+
+- Drop support for v1.0.0
+- v2.0.0 redesigned the API of SafeOperation:
+
+  SafeOperation.run
+
+  SafeOperation#value_or
+  SafeOperation#value_or_else
+  SafeOperation#and_then
+  SafeOperation#or_else
+
 ## 1.0.0
 
 First implementation ([#1][pr1]) - 2017.07.09